refactor configure file loader for geo loader and v5
