+++ infra/conf/cfgcommon/common.go
@@ -1,4 +1,4 @@
-package conf
+package cfgcommon
 
 import (
 	"encoding/json"
@@ -8,6 +8,8 @@ import (
 	"github.com/v2fly/v2ray-core/v4/common/net"
 	"github.com/v2fly/v2ray-core/v4/common/protocol"
 )
+
+//go:generate go run github.com/v2fly/v2ray-core/v4/common/errors/errorgen
 
 type StringList []string
