+++ infra/conf/cfgcommon/errors.generated.go
@@ -1,0 +1,9 @@
+package cfgcommon
+
+import "github.com/v2fly/v2ray-core/v4/common/errors"
+
+type errPathObjHolder struct{}
+
+func newError(values ...interface{}) *errors.Error {
+	return errors.New(values...).WithPathObj(errPathObjHolder{})
+}
