+++ infra/conf/dns_proxy.go
@@ -2,15 +2,16 @@ package conf
 
 import (
 	"github.com/golang/protobuf/proto"
+	"github.com/v2fly/v2ray-core/v4/infra/conf/cfgcommon"
 
 	"github.com/v2fly/v2ray-core/v4/common/net"
 	"github.com/v2fly/v2ray-core/v4/proxy/dns"
 )
 
 type DNSOutboundConfig struct {
-	Network Network  `json:"network"`
-	Address *Address `json:"address"`
-	Port    uint16   `json:"port"`
+	Network cfgcommon.Network  `json:"network"`
+	Address *cfgcommon.Address `json:"address"`
+	Port    uint16             `json:"port"`
 }
 
 func (c *DNSOutboundConfig) Build() (proto.Message, error) {
