+++ main/distro/all/all.go
@@ -67,6 +67,9 @@ import (
 	_ "github.com/v2fly/v2ray-core/v4/transport/internet/headers/wechat"
 	_ "github.com/v2fly/v2ray-core/v4/transport/internet/headers/wireguard"
 
+	// Geo loaders
+	_ "github.com/v2fly/v2ray-core/v4/infra/conf/geodata/standard"
+
 	// JSON config support. Choose only one from the two below.
 	// The following line loads JSON from v2ctl
 	// _ "github.com/v2fly/v2ray-core/v4/main/json"
